requirements.txt
# Core dependencies for ProHance Workflow Process Generator
langchain>=0.3.27,<0.4.0
langchain-community>=0.3.29,<0.4.0
langchain-openai>=0.3.32,<0.4.0
langgraph>=0.6.6,<0.7.0
langchain-core>=0.3.75,<0.4.0
pydantic>=2.11.7,<3.0.0
python-dotenv>=0.9.9
gradio>=4.44.1
langfuse>=3.3.3,<4.0.0

# Additional dependencies for web hosting
uvicorn>=0.35.0
fastapi>=0.116.0
requests>=2.32.0
pyyaml>=6.0.0